requirements.txt
fastapi==0.115.12
SQLAlchemy==2.0.39
pydantic==2.10.6
email-validator==2.2.0
python-jose==3.4.0
redis==5.2.1
uvicorn==0.34.0
mysql-connector-python==9.2.0
